Simplified physics calculations and generalized Entity update parameters
Instead of calculating the time that has passed since last frame, in every
frame, to get the time delta for physics calculations, a fixed constant
`SECONDS_PER_FRAME` is now used. The timed number was practically the same
as the new constant anyway. This makes the `Level.update` code much simpler.

Additionally, the arguments of the `Entity.update` method have been changed.
It now expects a `Level` object instead of a `Decor` object. This
prepares for the future, where entities will have to see not only the
level's decor, but also the other entities in the level (e.g., the player
will have to see the enemies to detect collision).